requirements.txt
anovelmous-grammar==0.1.3
Django==1.8.2
django-filter==0.10.0
django-rest-swagger==0.3.2
djangorestframework==3.1.3
djangorestframework-filters==0.3.0
drf-extensions==0.2.7
drf-nested-routers==0.9.0
Markdown==2.6.2
nltk==3.0.3
numpy==1.9.2
psycopg2==2.5.4
PyYAML==3.11
redis==2.10.3
requests==2.7.0
